Move unit tests to their own folder.
